Merge bundle and crossplane resources charts into this repository (#175)
* Merge bundle and crossplane resources charts into this repository

With this, the Crossplane resources, bundle and the WC app charts are merged into the same repo, using 2 different charts instead of 3.

## Structure

The main idea is that there's an bundle chart, which is installed on the MC and installs the necessary Crossplane resources and the WC chart.

The WC chart is now installed as a Flux `HelmRelease`.

The main chart is named `aws-load-balancer-controller-bundle` and the WC chart is `aws-load-balancer-controller-driver`.

## Chart values

To simplify installation, all the values intended for the actual WC chart are provided as is to the App chart, and will be passed over to the ConfigMap for the WC chart. This implies that the `values.yaml` and `values.schema.json` files are duplicated in both charts.

## Automation

The WC app chart is pushed to the `giantswarm` catalog, while the bundle chart is pushed to both the `giantswarm` and the `control-plane` catalogs, as well as the capa app collection.

## Versioning

Both charts share the same release cycle and the same versions.

## Migration

The migration path is to remove the old App and install the new one. Details will be added to the release notes and the README

Towards https://github.com/giantswarm/giantswarm/issues/34551

* Create namespace on install

* Move polex to policy-exceptions NS

* Add managed-by tag in crossplane resource

* Fix template name